Force validator restore to the dashboard state. vtrust and trust should be less wonky now
