+++ meta/meta.json
@@ -1,3 +1,3 @@
 {
-  "subnet_version": "2.6.8"
+  "subnet_version": "2.7.0"
 }
